Handle fractional quantities and keep the merchant's restock choice
Stores can allow decimal quantities through
woocommerce_quantity_input_step_admin, and parseInt turned 0.5 into 0, so
the option never became available for them. The template truncated
fractional reduced stock the same way.

Unticking is also no longer forgotten when quantities return to zero. The
choice now lasts as long as the refund panel does, which matches how it
already survived cancelling and reopening.

Both raised in review on #68230.

+++ plugins/woocommerce/includes/admin/meta-boxes/views/html-order-items.php
@@ -38,7 +38,7 @@ $restockable_items = array();
 foreach ( $line_items as $restock_item_id => $restock_item ) {
 	$restock_product = $restock_item->get_product();
 
-	if ( $restock_product && $restock_product->managing_stock() && (int) $restock_item->get_meta( '_reduced_stock', true ) > 0 ) {
+	if ( $restock_product && $restock_product->managing_stock() && (float) $restock_item->get_meta( '_reduced_stock', true ) > 0 ) {
 		$restockable_items[ $restock_item_id ] = true;
 	}
 }
